Wk 3 - Duel - R6 Double KO Fixed
I saw that I only got double KOs for the fights so I decided to check
out the damage code. I found that it is necessary to have the break at
the end of the winnerCheck FN so that round 6 does not repeat itself
and calculate a ridiculous amount of damage in the end.